requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
beautifulsoup4==4.9.3
bleach==3.3.0
certifi==2020.11.8
chardet==3.0.4
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
distlib==0.3.1
Django==3.1.4
django-cors==0.1
django-extensions==3.1.1
django-filter==2.4.0
django-rest-swagger==2.2.0
django-swagger-render==0.1.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
docutils==0.17.1
filelock==3.0.12
google==3.0.0
idna==2.10
importlib-metadata==4.0.1
isort==5.6.4
itypes==1.2.0
Jinja2==2.11.3
joblib==1.0.1
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.3.4
mccabe==0.6.1
numpy==1.20.1
openapi-codec==1.3.2
packaging==20.9
pandas==1.2.2
Pillow==8.1.0
Pygments==2.8.1
PyJWT==1.7.1
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.4
pytz==2020.5
pywin32-ctypes==0.2.0
PyYAML==5.4.1
requests==2.25.0
scikit-learn==0.24.1
scipy==1.6.1
selenium==3.141.0
simplejson==3.17.2
six==1.15.0
sklearn==0.0
soupsieve==2.1
sqlparse==0.4.1
threadpoolctl==2.1.0
toml==0.10.2
tqdm==4.60.0
uritemplate==3.0.1
urllib3==1.26.2
virtualenv==20.4.4
webencodings==0.5.1
wrapt==1.12.1
zipp==3.4.1
